requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
openai==1.3.0
sqlalchemy==2.0.23
pydantic==2.5.2
python-multipart==0.0.6
pymysql==1.1.0
cryptography==41.0.5
boto3==1.34.49
python-jose==3.3.0
pytest==7.4.3
httpx==0.25.1
alembic==1.12.1 